GH-4009: Check tests. Reorganise test grouping.
