+++ jena-arq/src/test/java/org/apache/jena/rdf12/TestNTriplesTripleTerms.java
@@ -19,7 +19,7 @@
  *   SPDX-License-Identifier: Apache-2.0
  */
 
-package org.apache.jena.rdf12.parse;
+package org.apache.jena.rdf12;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
@@ -28,60 +28,39 @@ import org.junit.jupiter.api.Test;
 import org.apache.jena.riot.Lang;
 import org.apache.jena.riot.RDFParser;
 import org.apache.jena.riot.RiotException;
-import org.apache.jena.riot.lang.turtlejcc.TurtleJCC;
 import org.apache.jena.riot.system.ErrorHandler;
 import org.apache.jena.riot.system.ErrorHandlerFactory;
 import org.apache.jena.riot.system.StreamRDF;
 import org.apache.jena.riot.system.StreamRDFLib;
 
 /**
- * Test parsing of occurrences and triple terms in Turtle.
+ * Basic parsing tests of triple terms
  */
-public class TestTurtleStarParse {
+public class TestNTriplesTripleTerms {
 
     private ErrorHandler silent = ErrorHandlerFactory.errorHandlerStrictNoLogging;
     private static StreamRDF sink = StreamRDFLib.sinkNull();
 
-    @Test public void parse_turtle_good_1()    { parse("<<:s :p :o>> :q 1 . "); }
+    @Test public void parse_nt_good_1()    { parse("<x:s> <x:q> <<( <x:s> <x:p> <x:o>)>> . "); }
 
-    @Test public void parse_turtle_good_2()    { parse(":x :p <<:s :p :o>> . "); }
+    @Test public void parse_nt_good_2()    { parse("_:b <x:p> <<(_:b <x:p> _:o)>>. "); }
 
-    @Test public void parse_turtle_good_3()    { parse(":x :p [ :q <<:s :p :o>> ]. "); }
-
-    @Test public void parse_turtle_good_4()    { parse("( <<:s :p :o>> ) :p :z . "); }
-
-    @Test public void parse_turtle_good_5()    { parse("( <<[] :p []>> ) :p :z . "); }
-
-    @Test public void parse_turtle_good_10()   { parse("<<:s :p <<:x :r :z >>>> :q 1 . "); }
-
-    @Test public void parse_turtle_good_20()   { parse(":a :p <<:s :p <<:x :r :z >>>> . "); }
-
-    // Reified triple declarations.
-    public void parse_turtle_good_30()           { parse("<<:s :p :o>> . "); }
-
-    // Triple Terms.
-    @Test public void parse_turtle_good_50()   { parse(":a :p <<(:s :p <<(:x :r :z )>>)>> . "); }
+    @Test public void parse_nt_good_3()    { parse("<x:x> <x:y> <<(<x:s1> <x:p1> <<( <x:s> <x:p> '1' )>> )>> ."); }
 
     @Test
-    public void parse_turtle_bad_2()           { parseException("<<(:s :p :o )>> :q 1 . "); }
+    public void parse_nt_bad_1()           { parseException("<< <x:s> <x:p> <x:o> >> . "); }
 
     @Test
-    public void parse_turtle_bad_3()           { parseException(":s :q <<:s :p (3) >> . "); }
+    public void parse_nt_bad_2()           { parseException("<<( <x:s> <x:p> <x:o>)>> <x:y> <x:z> . "); }
 
     @Test
-    public void parse_turtle_bad_4()           { parseException(":s :p <<( 3 :q :t >> . "); }
+    public void parse_nt_bad_3()           { parseException("<x:s> <<( <x:s> <x:p> <x:o>)>> <x:x>  . "); }
 
     private void parseException(String string) {
         assertThrows(RiotException.class, ()->parse(string));
     }
 
     private void parse(String string) {
-        string = "PREFIX : <http://example/>\n"+string;
-
-        Lang lang1 = Lang.TURTLE;
-        Lang lang2 = TurtleJCC.TTLJCC;
-        Lang lang = lang1;
-
-        RDFParser.fromString(string, lang).errorHandler(silent).parse(sink);
+        RDFParser.fromString(string, Lang.NTRIPLES).errorHandler(silent).parse(sink);
     }
 }
