+++ jena-arq/src/test/java/org/apache/jena/rdf12/TestTrigParseTripleTerms.java
@@ -19,7 +19,7 @@
  *   SPDX-License-Identifier: Apache-2.0
  */
 
-package org.apache.jena.rdf12.parse;
+package org.apache.jena.rdf12;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
@@ -33,34 +33,45 @@ import org.apache.jena.riot.system.ErrorHandlerFactory;
 import org.apache.jena.riot.system.StreamRDF;
 import org.apache.jena.riot.system.StreamRDFLib;
 
+
 /**
- * Basic parsing tests of triple terms
+ * Test parsing of triple term constructs for TriG.
  */
-public class TestNTriplesTripleTerms {
+public class TestTrigParseTripleTerms {
 
     private ErrorHandler silent = ErrorHandlerFactory.errorHandlerStrictNoLogging;
     private static StreamRDF sink = StreamRDFLib.sinkNull();
 
-    @Test public void parse_nt_good_1()    { parse("<x:s> <x:q> <<( <x:s> <x:p> <x:o>)>> . "); }
+    @Test public void parse_trig_good_tripleTerm_1()    { parse(":x :q <<(:s :p :o)>>  . "); }
 
-    @Test public void parse_nt_good_2()    { parse("_:b <x:p> <<(_:b <x:p> _:o)>>. "); }
+    @Test public void parse_trig_good_tripleTerm_2()    { parse(":x :p <<( :s :p <<( :Ls1 :1p :o1 )>> )>> . "); }
 
-    @Test public void parse_nt_good_3()    { parse("<x:x> <x:y> <<(<x:s1> <x:p1> <<( <x:s> <x:p> '1' )>> )>> ."); }
+    @Test public void parse_trig_good_reifiedTriple_1()    { parse("<<:s :p :o>> :q 1 . "); }
+
+    @Test public void parse_trig_good_reifiedTriple_2()    { parse(":x :p <<:s :p :o>> . "); }
+
+    @Test public void parse_trig_good_reifiedTriple_3()    { parse(":x :p [ :q <<:s :p :o>> ]. "); }
+
+    @Test public void parse_trig_good_reifiedTriple_4()    { parse("( <<:s :p :o>> ) :p :z . "); }
+
+    @Test public void parse_trig_good_reifiedTriple_5()    { parse("<<:s :p :o>> . "); }
+
+    @Test public void parse_trig_good_reifiedTriple_10()   { parse("<<:s :p <<:x :r :z >>>> :q 1 . "); }
+
+    @Test public void parse_trig_good_reifiedTriple_20()   { parse(":a :p <<:s :p <<:x :r :z >>>> . "); }
 
     @Test
-    public void parse_nt_bad_1()           { parseException("<< <x:s> <x:p> <x:o> >> . "); }
+    public void parse_trig_bad_1()           { parseException("GRAPH <<(:s :p :o)>> {} "); }
 
     @Test
-    public void parse_nt_bad_2()           { parseException("<<( <x:s> <x:p> <x:o>)>> <x:y> <x:z> . "); }
-
-    @Test
-    public void parse_nt_bad_3()           { parseException("<x:s> <<( <x:s> <x:p> <x:o>)>> <x:x>  . "); }
+    public void parse_trig_bad_2()           { parseException("<<(:s :p :o)>> { :s :p :o } "); }
 
     private void parseException(String string) {
         assertThrows(RiotException.class, ()->parse(string));
     }
 
     private void parse(String string) {
-        RDFParser.fromString(string, Lang.NTRIPLES).errorHandler(silent).parse(sink);
+        string = "PREFIX : <http://example/>\n"+string;
+        RDFParser.fromString(string, Lang.TRIG).errorHandler(silent).parse(sink);
     }
 }
