+++ jena-arq/src/test/java/org/apache/jena/rdf12/TestTurtleTripleTermsParse.java
@@ -19,7 +19,7 @@
  *   SPDX-License-Identifier: Apache-2.0
  */
 
-package org.apache.jena.rdf12.parse;
+package org.apache.jena.rdf12;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
@@ -28,46 +28,60 @@ import org.junit.jupiter.api.Test;
 import org.apache.jena.riot.Lang;
 import org.apache.jena.riot.RDFParser;
 import org.apache.jena.riot.RiotException;
+import org.apache.jena.riot.lang.turtlejcc.TurtleJCC;
 import org.apache.jena.riot.system.ErrorHandler;
 import org.apache.jena.riot.system.ErrorHandlerFactory;
 import org.apache.jena.riot.system.StreamRDF;
 import org.apache.jena.riot.system.StreamRDFLib;
 
 /**
- * Basic parsing tests of RDF triple term constructs for N-Quads
+ * Test parsing of occurrences and triple terms in Turtle.
  */
-public class TestNQuadsTripleTerms {
+public class TestTurtleTripleTermsParse {
 
     private ErrorHandler silent = ErrorHandlerFactory.errorHandlerStrictNoLogging;
     private static StreamRDF sink = StreamRDFLib.sinkNull();
 
-    @Test public void parse_nq_good_1()    { parse("<x:s> <x:q> <<( <x:s> <x:p> <x:o> )>> . "); }
+    @Test public void parse_turtle_good_1()    { parse("<<:s :p :o>> :q 1 . "); }
 
-    @Test public void parse_nq_good_2()    { parse("<x:s> <x:p> <<( <x:s> <x:p> <x:o> )>> <http://example/g> . "); }
+    @Test public void parse_turtle_good_2()    { parse(":x :p <<:s :p :o>> . "); }
 
-    @Test public void parse_nq_good_3()    { parse("<x:s> <x:q> <<( <x:s> <x:p> <<(<x:s1> <x:p1> <x:o1>)>> )>> . "); }
+    @Test public void parse_turtle_good_3()    { parse(":x :p [ :q <<:s :p :o>> ]. "); }
 
-    @Test public void parse_nq_good_4()    { parse("<x:s> <x:p> <<( <x:s> <x:p> <<(<x:s1> <x:p1> <x:o1>)>> )>> <http://example/g> . "); }
+    @Test public void parse_turtle_good_4()    { parse("( <<:s :p :o>> ) :p :z . "); }
 
-    @Test public void parse_nq_good_5()    { parse("_:b <x:p> <<(_:b <x:p> _:o)>> _:g . "); }
+    @Test public void parse_turtle_good_5()    { parse("( <<[] :p []>> ) :p :z . "); }
+
+    @Test public void parse_turtle_good_10()   { parse("<<:s :p <<:x :r :z >>>> :q 1 . "); }
+
+    @Test public void parse_turtle_good_20()   { parse(":a :p <<:s :p <<:x :r :z >>>> . "); }
+
+    // Reified triple declarations.
+    public void parse_turtle_good_30()           { parse("<<:s :p :o>> . "); }
+
+    // Triple Terms.
+    @Test public void parse_turtle_good_50()   { parse(":a :p <<(:s :p <<(:x :r :z )>>)>> . "); }
 
     @Test
-    public void parse_nq_bad_1()           { parseException("<<<x:s> <x:p> <x:o>>> <x:q> <x:z> . "); }
+    public void parse_turtle_bad_2()           { parseException("<<(:s :p :o )>> :q 1 . "); }
 
     @Test
-    public void parse_nq_bad_2()           { parseException("<<(<x:s> <x:p> <x:o>)>> <x:q> 'ABC' . "); }
+    public void parse_turtle_bad_3()           { parseException(":s :q <<:s :p (3) >> . "); }
 
     @Test
-    public void parse_nq_bad_3()           { parseException("<x:s> <<(<x:s> <x:p> <x:o>)>> <x:o> <http://example/g> . "); }
-
-    @Test
-    public void parse_nq_bad_4()           { parseException("<x:s> <x:p> <x:o> <<(<x:gs> <x:gp> <x:go> )>> . "); }
+    public void parse_turtle_bad_4()           { parseException(":s :p <<( 3 :q :t >> . "); }
 
     private void parseException(String string) {
         assertThrows(RiotException.class, ()->parse(string));
     }
 
     private void parse(String string) {
-        RDFParser.fromString(string, Lang.NQUADS).errorHandler(silent).parse(sink);
+        string = "PREFIX : <http://example/>\n"+string;
+
+        Lang lang1 = Lang.TURTLE;
+        Lang lang2 = TurtleJCC.TTLJCC;
+        Lang lang = lang1;
+
+        RDFParser.fromString(string, lang).errorHandler(silent).parse(sink);
     }
 }
