+++ jena-arq/src/test/java/org/apache/jena/system/TC_System.java
@@ -19,19 +19,41 @@
  *   SPDX-License-Identifier: Apache-2.0
  */
 
-package org.apache.jena.rdf12.parse;
+package org.apache.jena.system;
 
 import org.junit.platform.suite.api.SelectClasses;
 import org.junit.platform.suite.api.Suite;
 
+import org.apache.jena.http.auth.TS_HttpAuth;
+import org.apache.jena.rdfs.TS_InfRdfs;
+import org.apache.jena.sparql.graph.TS_Graph;
+import org.apache.jena.sparql.sse.TS_SSE;
+import org.apache.jena.sparql.transaction.TS_Transaction;
+import org.apache.jena.sparql.util.TS_DyadicDatasetGraphs;
+import org.apache.jena.sparql.util.TS_Util;
+import org.apache.jena.sparql.util.compose.TS_DatasetCollectors;
+import org.apache.jena.sparql.util.iso.TS_Iso;
+import org.apache.jena.system.buffering.TS_Buffering;
+import org.apache.jena.util.TS_UtilsARQ;
+
 @Suite
 @SelectClasses({
-    // ** To be removed?
-    TestNQuadsTripleTerms.class,
-    TestNTriplesTripleTerms.class,
-    TestTrigParseTripleTerms.class,
-    TestTurtleStarParse.class,
-    TestSPARQL12Parse.class
+    TS_System.class
+    , TS_SSE.class
+    , TS_Graph.class
+    , TS_DyadicDatasetGraphs.class
+    , TS_DatasetCollectors.class
+    , TS_Util.class
+    , TS_Iso.class
+    , TS_Transaction.class
+
+    , TS_UtilsARQ.class
+    , TS_System.class
+    , TS_Buffering.class
+
+    , TS_InfRdfs.class
+    , TS_HttpAuth.class
 })
-public class TS_RDFStar_Parse {
-}
+
+public class TC_System {}
+
